requirements.txt
alembic==1.0.9
aniso8601==6.0.0
appdirs==1.4.3
attrs==19.1.0
beautifulsoup4==4.7.1
black==19.3b0
Click==7.0
Flask==1.0.2
flask-marshmallow==0.10.0
Flask-Migrate==2.4.0
flask-restplus==0.12.1
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
lxml==4.3.3
Mako==1.0.9
MarkupSafe==1.1.1
marshmallow==3.0.0rc4
marshmallow-sqlalchemy==0.16.2
psycopg2-binary==2.8.2
pyrsistent==0.14.11
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
pytz==2019.1
redis==3.2.1
rq==1.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9.1
SQLAlchemy==1.3.3
toml==0.10.0
urllib3==1.24.2
webargs==5.3.0
Werkzeug==0.15.2
